+++ Src/Sla/DECCORE/Rule2Comp2Mult.cs
@@ -26,7 +26,7 @@ namespace Sla.DECCORE
         {
             data.opSetOpcode(op, OpCode.CPUI_INT_MULT);
             int size = op.getIn(0).getSize();
-            Varnode negone = data.newConstant(size, Globals.calc_mask(size));
+            Varnode negone = data.newConstant(size, Globals.calc_mask((uint)size));
             data.opInsertInput(op, negone, 1);
             return 1;
         }
