+++ Src/Sla/DECCORE/RuleMultNegOne.cs
@@ -28,7 +28,7 @@ namespace Sla.DECCORE
             Varnode constvn = op.getIn(1) ?? throw new ApplicationException();
 
             if (!constvn.isConstant()) return 0;
-            if (constvn.getOffset() != Globals.calc_mask(constvn.getSize())) return 0;
+            if (constvn.getOffset() != Globals.calc_mask((uint)constvn.getSize())) return 0;
 
             data.opSetOpcode(op, OpCode.CPUI_INT_2COMP);
             data.opRemoveInput(op, 1);
